requirements.txt
google-generativeai
exa-py
python-dotenv
beautifulsoup4
requests
